Safari-611.1.10/LayoutTests/fast/selectors/matches-scope-expected.txt

This test makes sure that :scope works correctly with the matches() API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS theTarget.matches(":scope") is true
PASS theTarget.matches(":not(:scope)") is false
PASS theTarget.matches("body :scope") is true
PASS theTarget.matches("body>:scope") is true
PASS theTarget.matches("sibling + :scope") is true
PASS theTarget.matches("sibling ~ :scope") is true
PASS theTarget.matches("#theTarget:scope") is true
PASS theTarget.matches(":scope#theTarget") is true
PASS theTarget.matches("[webkit]:scope#theTarget") is true
PASS theTarget.matches(":not([webkit=fast]):scope#theTarget") is false
PASS theTarget.matches(":scope target") is false
PASS theTarget.matches(":scope > target") is false
PASS theTarget.matches(":scope + target") is false
PASS theTarget.matches(":scope ~ target") is false
PASS theTarget.matches(":scope *") is false
PASS theTarget.matches(":scope > *") is false
PASS theTarget.matches(":scope + *") is false
PASS theTarget.matches(":scope ~ *") is false
PASS successfullyParsed is true

TEST COMPLETE
